Remove simplify_ops and always compile in release mode
