+++ charon/src/charon-driver.rs
@@ -65,7 +65,6 @@ mod remove_read_discriminant;
 mod remove_unused_locals;
 mod reorder_decls;
 mod rust_to_local_ids;
-mod simplify_ops;
 mod translate_functions_to_ullbc;
 mod translate_types;
 mod types;
