+++ charon/src/driver.rs
@@ -15,7 +15,6 @@ use crate::remove_read_discriminant;
 use crate::remove_unused_locals;
 use crate::reorder_decls;
 use crate::rust_to_local_ids;
-use crate::simplify_ops;
 use crate::translate_functions_to_ullbc;
 use crate::translate_types;
 use crate::ullbc_to_llbc;
@@ -254,18 +253,7 @@ pub fn translate(sess: &Session, tcx: TyCtxt, internal: &CharonCallbacks) -> Res
             &ullbc_globals,
         );
 
-        // # Step 9: simplify the calls to unops and binops
-        // Note that we assume that the sequences have been flattened.
-        simplify_ops::simplify(options.release, &fmt_ctx, &mut llbc_funs, &mut llbc_globals);
-
-        for def in &llbc_funs {
-            trace!(
-                "# After binop simplification:\n{}\n",
-                def.fmt_with_decls(&type_defs, &llbc_funs, &llbc_globals)
-            );
-        }
-
-        // # Step 10: reconstruct the asserts
+        // # Step 9: reconstruct the asserts
         reconstruct_asserts::transform(&fmt_ctx, &mut llbc_funs, &mut llbc_globals);
 
         for def in &llbc_funs {
@@ -275,10 +263,10 @@ pub fn translate(sess: &Session, tcx: TyCtxt, internal: &CharonCallbacks) -> Res
             );
         }
 
-        // # Step 11: Remove the discriminant reads (merge them with the switches)
+        // # Step 10: Remove the discriminant reads (merge them with the switches)
         remove_read_discriminant::transform(&fmt_ctx, &mut llbc_funs, &mut llbc_globals);
 
-        // # Step 12: add the missing assignments to the return value.
+        // # Step 11: add the missing assignments to the return value.
         // When the function return type is unit, the generated MIR doesn't
         // set the return value to `()`. This can be a concern: in the case
         // of Aeneas, it means the return variable contains ⊥ upon returning.
@@ -288,15 +276,15 @@ pub fn translate(sess: &Session, tcx: TyCtxt, internal: &CharonCallbacks) -> Res
         // the main or at compile-time).
         insert_assign_return_unit::transform(&fmt_ctx, &mut llbc_funs, &mut llbc_globals);
 
-        // # Step 13: remove the drops of locals whose type is `Never` (`!`). This
+        // # Step 12: remove the drops of locals whose type is `Never` (`!`). This
         // is in preparation of the next transformation.
         remove_drop_never::transform(&fmt_ctx, &mut llbc_funs, &mut llbc_globals);
 
-        // # Step 14: remove the locals which are never used. After doing so, we
+        // # Step 13: remove the locals which are never used. After doing so, we
         // check that there are no remaining locals with type `Never`.
         remove_unused_locals::transform(&fmt_ctx, &mut llbc_funs, &mut llbc_globals);
 
-        // # Step 15: compute which functions are potentially divergent. A function
+        // # Step 14: compute which functions are potentially divergent. A function
         // is potentially divergent if it is recursive, contains a loop or transitively
         // calls a potentially divergent function.
         // Note that in the future, we may complement this basic analysis with a
@@ -305,7 +293,7 @@ pub fn translate(sess: &Session, tcx: TyCtxt, internal: &CharonCallbacks) -> Res
         // Because we don't have loops, constants are not yet touched.
         let _divergent = divergent::compute_divergent_functions(&ordered_decls, &llbc_funs);
 
-        // # Step 16: generate the files.
+        // # Step 15: generate the files.
         export::export_llbc(
             crate_name,
             &ordered_decls,
