+++ tests/Makefile
@@ -56,11 +56,8 @@ test-%:
 	$(CHARON_CMD) --dest $(DEST)/llbc
 	$(CHARON_CMD) --dest $(DEST)/ullbc --ullbc
 	$(CHARON_CMD) --dest $(DEST)/llbc_prom --mir_promoted
-	$(CHARON_CMD) --dest $(DEST)/llbc_opt --mir_optimized
-	$(CHARON_CMD) --dest $(DEST)/llbc_release --release
-	$(CHARON_CMD) --dest $(DEST)/llbc_release_prom --release --mir_promoted
 # TODO: this fails for now (there is some very low-level desugaring happening)
-#	$(CHARON_CMD) --dest $(DEST)/llbc_release_opt --release --mir_optimized 
+#	$(CHARON_CMD) --dest $(DEST)/llbc_opt --mir_optimized 
 
 else
